+++ vitest.config.ts
@@ -3,6 +3,8 @@ import { defineConfig, defineProject } from 'vitest/config'
 
 export default defineConfig({
   test: {
+    globals: true,
+    reporters: 'dot',
     projects: [
       // type tests using vitest typecheck
       defineProject({